review changes: refacto radio buttons
